+++ docs/CHANGELOG.md
@@ -2,6 +2,8 @@
 > windowsboy111
 
 ## [Unreleased]
+
+## [2.1.1]
 ### Added
 - `/avatar` command
 - `/res` command to see CPU / RAM usages
@@ -14,6 +16,8 @@
 - no attribute hint
 - stage is now stable
 - member dm welcome
+- spam detecion system now won't crash randomly cannot delete the message
+- `@Merlin` prefix now works
 
 ## [2.1.0]
 ### Added
@@ -143,7 +147,8 @@
 - no more verbose mode :O
 
 
-[Unreleased]:   https://github.com/windowsboy111/Merlin-py/compare/r2.1.0...HEAD
+[Unreleased]:   https://github.com/windowsboy111/Merlin-py/compare/2.1.1...HEAD
+[2.1.1]:        https://github.com/windowsboy111/Merlin-py/compare/r2.1.0...2.1.1
 [2.1.0]:        https://github.com/windowsboy111/Merlin-py/compare/2.1-b3...r2.1.0
 [2.1-b3]:       https://github.com/windowsboy111/Merlin-py/compare/2.1-b2...2.1-b3
 [2.1-b2]:       https://github.com/windowsboy111/Merlin-py/compare/2.1-b1...2.1-b2